api_wallet/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
python-dotenv==1.0.1
pytest-django==4.5.2
psycopg2-binary==2.9.3